Try to reach a high level
Ahhh, the 80s. That totally awesome decade that brought us big hair, 24/7 music videos, the Brat Pack and Simon. And now, the classic electronic game of sharp memories and fast hands is back to exercise the minds of a new generation while giving the older fogies a little neuron nostalgia. If you're new to the game, here's how it works: just activate the device, watch and listen as the four buttons light up randomly and produce different sounds. Then, press the buttons in the same order to progress to the next round. With each round, the number of buttons to press increases as does the challenge.

Brain Health Expert Says:
"The hippocampus, which functions in memory storage and retrieval, is activated as you memorize, then replicate the sequence of lights and sounds."